+++ tests/shared/api.test.ts
@@ -5,7 +5,7 @@ const pangu = new Pangu();
 
 describe('API', () => {
   describe('spacingText()', () => {
-    it('should spacing text', () => {
+    it('spacing text', () => {
       // prettier-ignore
       expect(pangu.spacingText('聽說Hadoop工程師睡不著的時候都會MapReduce羊'))
                          .toBe('聽說 Hadoop 工程師睡不著的時候都會 MapReduce 羊');
@@ -17,7 +17,7 @@ describe('API', () => {
   });
 
   describe('hasProperSpacing()', () => {
-    it('should detect proper spacing', () => {
+    it('detect proper spacing', () => {
       expect(pangu.hasProperSpacing('♫ 每條大街小巷，每個工程師的嘴裡，見面第一句話，就是不要在過年前 Deploy ♫')).toBe(true);
       expect(pangu.hasProperSpacing('♫每條大街小巷，每個工程師的嘴裡，見面第一句話，就是不要在過年前Deploy♫')).toBe(false);
     });