+++ tests/shared/cjk-alphabets-numbers.test.ts
@@ -3,8 +3,8 @@ import { describe, it, expect } from 'vitest';
 
 const pangu = new Pangu();
 
-describe('CJK A N 兩邊都加空格', () => {
-  it('should handle short text', () => {
+describe('CJK A N add space on both sides', () => {
+  it('handle short text', () => {
     expect(pangu.spacingText('中a')).toBe('中 a');
     expect(pangu.spacingText('a中')).toBe('a 中');
     expect(pangu.spacingText('1中')).toBe('1 中');
@@ -15,85 +15,85 @@ describe('CJK A N 兩邊都加空格', () => {
     expect(pangu.spacingText('1中a')).toBe('1 中 a');
   });
 
-  it('should handle alphabets', () => {
+  it('handle alphabets', () => {
     expect(pangu.spacingText('中文abc')).toBe('中文 abc');
     expect(pangu.spacingText('abc中文')).toBe('abc 中文');
   });
 
-  it('should handle numbers', () => {
+  it('handle numbers', () => {
     expect(pangu.spacingText('中文123')).toBe('中文 123');
     expect(pangu.spacingText('123中文')).toBe('123 中文');
   });
 
   // https://symbl.cc/en/unicode-table/#latin-1-supplement
-  it('should handle Latin-1 Supplement', () => {
+  it('handle Latin-1 Supplement', () => {
     expect(pangu.spacingText('中文Ø漢字')).toBe('中文 Ø 漢字');
     expect(pangu.spacingText('中文 Ø 漢字')).toBe('中文 Ø 漢字');
   });
 
   // https://symbl.cc/en/unicode-table/#greek-coptic
-  it('should handle Greek and Coptic', () => {
+  it('handle Greek and Coptic', () => {
     expect(pangu.spacingText('中文β漢字')).toBe('中文 β 漢字');
     expect(pangu.spacingText('中文 β 漢字')).toBe('中文 β 漢字');
     expect(pangu.spacingText('我是α，我是Ω')).toBe('我是 α，我是 Ω');
   });
 
   // https://symbl.cc/en/unicode-table/#number-forms
-  it('should handle Number Forms', () => {
+  it('handle Number Forms', () => {
     expect(pangu.spacingText('中文Ⅶ漢字')).toBe('中文 Ⅶ 漢字');
     expect(pangu.spacingText('中文 Ⅶ 漢字')).toBe('中文 Ⅶ 漢字');
   });
 
   // https://symbl.cc/en/unicode-table/#cjk-radicals-supplement
-  it('should handle CJK Radicals Supplement', () => {
+  it('handle CJK Radicals Supplement', () => {
     expect(pangu.spacingText('abc⻤123')).toBe('abc ⻤ 123');
     expect(pangu.spacingText('abc ⻤ 123')).toBe('abc ⻤ 123');
   });
 
   // https://symbl.cc/en/unicode-table/#kangxi-radicals
-  it('should handle Kangxi Radicals', () => {
+  it('handle Kangxi Radicals', () => {
     expect(pangu.spacingText('abc⾗123')).toBe('abc ⾗ 123');
     expect(pangu.spacingText('abc ⾗ 123')).toBe('abc ⾗ 123');
   });
 
   // https://symbl.cc/en/unicode-table/#hiragana
-  it('should handle Hiragana', () => {
+  it('handle Hiragana', () => {
     expect(pangu.spacingText('abcあ123')).toBe('abc あ 123');
     expect(pangu.spacingText('abc あ 123')).toBe('abc あ 123');
   });
 
   // https://symbl.cc/en/unicode-table/#katakana
-  it('should handle Katakana', () => {
+  it('handle Katakana', () => {
     expect(pangu.spacingText('abcア123')).toBe('abc ア 123');
     expect(pangu.spacingText('abc ア 123')).toBe('abc ア 123');
   });
 
   // https://symbl.cc/en/unicode-table/#bopomofo
-  it('should handle Bopomofo', () => {
+  it('handle Bopomofo', () => {
     expect(pangu.spacingText('abcㄅ123')).toBe('abc ㄅ 123');
     expect(pangu.spacingText('abc ㄅ 123')).toBe('abc ㄅ 123');
   });
 
   // https://symbl.cc/en/unicode-table/#enclosed-cjk-letters-and-months
-  it('should handle Enclosed CJK Letters And Months', () => {
+  it('handle Enclosed CJK Letters And Months', () => {
     expect(pangu.spacingText('abc㈱123')).toBe('abc ㈱ 123');
     expect(pangu.spacingText('abc ㈱ 123')).toBe('abc ㈱ 123');
   });
 
   // https://symbl.cc/en/unicode-table/#cjk-unified-ideographs-extension-a
-  it('should handle CJK Unified Ideographs Extension-A', () => {
+  it('handle CJK Unified Ideographs Extension-A', () => {
     expect(pangu.spacingText('abc㐂123')).toBe('abc 㐂 123');
     expect(pangu.spacingText('abc 㐂 123')).toBe('abc 㐂 123');
   });
 
   // https://symbl.cc/en/unicode-table/#cjk-unified-ideographs
-  it('should handle CJK Unified Ideographs', () => {
+  it('handle CJK Unified Ideographs', () => {
     expect(pangu.spacingText('abc丁123')).toBe('abc 丁 123');
     expect(pangu.spacingText('abc 丁 123')).toBe('abc 丁 123');
   });
 
   // https://symbl.cc/en/unicode-table/#cjk-compatibility-ideographs
-  it('should handle CJK Compatibility Ideographs', () => {
+  it('handle CJK Compatibility Ideographs', () => {
     expect(pangu.spacingText('abc車123')).toBe('abc 車 123');
     expect(pangu.spacingText('abc 車 123')).toBe('abc 車 123');
   });